+++ libs/designsystem/checkbox/src/checkbox.component.integration.spec.ts
