+++ libs/designsystem/checkbox/src/checkbox.component.scss
