+++ libs/designsystem/checkbox/src/checkbox.component.spec.ts
