+++ libs/designsystem/src/lib/index.ts
@@ -21,6 +21,7 @@ export * from '@kirbydesign/designsystem/range';
 export * from '@kirbydesign/designsystem/popover';
 export * from '@kirbydesign/designsystem/loading-overlay';
 export * from '@kirbydesign/designsystem/divider';
+export * from '@kirbydesign/designsystem/checkbox';
 
 export * from './scss/scss-helper';
