+++ libs/designsystem/src/lib/kirby.module.ts
@@ -27,6 +27,7 @@ import {
   LoadingOverlayService,
 } from '@kirbydesign/designsystem/loading-overlay';
 import { DividerComponent } from '@kirbydesign/designsystem/divider';
+import { CheckboxComponent } from '@kirbydesign/designsystem/checkbox';
 import { AccordionItemComponent } from './components/accordion/accordion-item.component';
 import { AccordionDirective } from './components/accordion/accordion.directive';
 import { AppModule } from './components/app/app.module';
@@ -34,7 +35,6 @@ import { AvatarComponent } from './components/avatar/avatar.component';
 import { ButtonComponent } from './components/button/button.component';
 import { CalendarComponent } from './components/calendar/calendar.component';
 import { ChartsModule } from './components/charts';
-import { CheckboxComponent } from './components/checkbox/checkbox.component';
 import { DataTableModule } from './components/data-table/data-table.module';
 import { FabSheetComponent } from './components/fab-sheet/fab-sheet.component';
 import { DateInputDirective } from './components/form-field/directives/date/date-input.directive';
