+++ libs/designsystem/tsconfig.json
@@ -23,7 +23,8 @@
       "@kirbydesign/designsystem/range": ["range/index.ts"],
       "@kirbydesign/designsystem/popover": ["popover/index.ts"],
       "@kirbydesign/designsystem/loading-overlay": ["loading-overlay/index.ts"],
-      "@kirbydesign/designsystem/divider": ["divider/index.ts"]
+      "@kirbydesign/designsystem/divider": ["divider/index.ts"],
+      "@kirbydesign/designsystem/checkbox": ["checkbox/index.ts"]
     },
     "esModuleInterop": true,
     "target": "es2020"
