+++ libs/designsystem/tsconfig.spec.json
@@ -26,7 +26,8 @@
       "@kirbydesign/designsystem/range": ["range/index.ts"],
       "@kirbydesign/designsystem/popover": ["popover/index.ts"],
       "@kirbydesign/designsystem/loading-overlay": ["loading-overlay/index.ts"],
-      "@kirbydesign/designsystem/divider": ["divider/index.ts"]
+      "@kirbydesign/designsystem/divider": ["divider/index.ts"],
+      "@kirbydesign/designsystem/checkbox": ["checkbox/index.ts"]
     }
   },
   "files": ["test.ts"],
